More improvments from PR feedbacks
