+++ packages/jsts/src/rules/plugin.ts
@@ -352,7 +352,7 @@ import { rule as S6627 } from './S6627';
 import { rule as S1607 } from './S1607';
 import { rule as S7060 } from './S7060';
 import type { Rule, Linter } from 'eslint';
-import { S2301 } from './S2301';
+import { rule as S2301 } from './S2301';
 import { rule as S7059 } from './S7059';
 
 export const rules: Record<string, Rule.RuleModule> = {
@@ -556,6 +556,7 @@ export const rules: Record<string, Rule.RuleModule> = {
   'no-return-type-any': S4324,
   'no-same-argument-assert': S5863,
   'no-same-line-conditional': S3972,
+  'no-selector-parameter': S2301,
   'no-self-compare': S6679,
   'no-self-import': S7060,
   'no-skipped-test': S1607,
@@ -683,7 +684,6 @@ export const rules: Record<string, Rule.RuleModule> = {
   'x-powered-by': S5689,
   'xml-parser-xxe': S2755,
   xpath: S4817,
-  s2301: S2301,
 };
 
 const recommendedLegacyConfig: Linter.Config = { plugins: ['sonarjs'], rules: {} };
